Make rewatch the default build system / cli (#7593)
* rescript_exe -> rescript_legacy_exe

* rescript.exe -> rescript-legacy.exe

* rewatch.exe -> rescript.exe

* cli/rescript.js -> cli/rescript-legacy.js

* cli/rewatch.js -> cli/rescript.js

* rewatch_exe -> rescript_exe

* rewatch.exe -> rescript.exe

* Remove rewatch/postinstall.js

* Rename runtime build scripts

* Fix gentype tests

* Build analysis tests with legacy build system

* Build playground with legacy build system

* Build more stuff with legacy build system

* CHANGELOG

* Testrepo: use pkg.pr.new

* Fix installation test

* Undo comment change
