+++ Makefile
@@ -17,7 +17,6 @@ dce:
 
 rewatch:
 	cargo build --manifest-path rewatch/Cargo.toml --release
-	cp rewatch/target/release/rewatch rewatch
 	./scripts/copyExes.js --rewatch
 
 ninja/ninja:
@@ -53,10 +52,10 @@ reanalyze:
 	reanalyze.exe -set-exit-code -all-cmt _build/default/compiler _build/default/tests -exclude-paths compiler/outcome_printer,compiler/ml,compiler/frontend,compiler/ext,compiler/depends,compiler/core,compiler/common,compiler/cmij,compiler/bsb_helper,compiler/bsb
 
 lib-bsb:
-	./scripts/buildRuntime.sh
+	./scripts/buildRuntimeLegacy.sh
 
 lib:
-	./scripts/buildRuntimeRewatch.sh
+	./scripts/buildRuntime.sh
 
 artifacts: lib
 	./scripts/npmPack.js --updateArtifactList
