+++ package.json
@@ -41,8 +41,8 @@
     "bsc": "cli/bsc.js",
     "bstracing": "cli/bstracing.js",
     "rescript": "cli/rescript.js",
-    "rescript-tools": "cli/rescript-tools.js",
-    "rewatch": "cli/rewatch.js"
+    "rescript-legacy": "cli/rescript-legacy.js",
+    "rescript-tools": "cli/rescript-tools.js"
   },
   "scripts": {
     "test": "node scripts/test.js -all",
