+++ packages/@rescript/darwin-arm64/bin.d.ts
@@ -3,11 +3,11 @@ export const binDir: string;
 export const binPaths: BinaryPaths;
 
 export type BinaryPaths = {
-  bsb_helper_exe: string,
-  bsc_exe: string,
-  ninja_exe: string,
-  rescript_exe: string,
-  rescript_tools_exe: string,
-  rescript_editor_analysis_exe: string,
-  rewatch_exe: string,
+  bsb_helper_exe: string;
+  bsc_exe: string;
+  ninja_exe: string;
+  rescript_legacy_exe: string;
+  rescript_tools_exe: string;
+  rescript_editor_analysis_exe: string;
+  rescript_exe: string;
 };
