+++ rewatch/testrepo/package.json
@@ -13,14 +13,14 @@
     ]
   },
   "dependencies": {
-    "rescript": "12.0.0-alpha.13"
+    "rescript": "https://pkg.pr.new/rescript-lang/rescript@6cac813"
   },
   "scripts": {
     "build": "../target/release/rewatch build .",
-    "build:rescript": "rescript build -with-deps",
+    "build:rescript": "rescript legacy build",
     "watch": "../target/release/rewatch watch .",
-    "watch:rescript": "rescript watch -with-deps",
+    "watch:rescript": "rescript legacy watch",
     "clean": "../target/release/rewatch clean .",
-    "clean:rescript": "rescript clean"
+    "clean:rescript": "rescript legacy clean"
   }
 }
