+++ scripts/copyExes.js
@@ -34,7 +34,7 @@ const shouldCopyNinja = args.values.all || args.values.ninja;
 const shouldCopyRewatch = args.values.all || args.values.rewatch;
 
 if (shouldCopyCompiler) {
-  copyExe(compilerBinDir, "rescript");
+  copyExe(compilerBinDir, "rescript-legacy");
   copyExe(compilerBinDir, "rescript-editor-analysis");
   copyExe(compilerBinDir, "rescript-tools");
   copyExe(compilerBinDir, "bsc");
@@ -46,17 +46,18 @@ if (shouldCopyNinja) {
 }
 
 if (shouldCopyRewatch) {
-  copyExe(rewatchDir, "rewatch");
+  copyExe(path.join(rewatchDir, "target", "release"), "rewatch", "rescript");
 }
 
 /**
  * @param {string} dir
  * @param {string} exe
+ * @param {string | undefined} renamed
  */
-function copyExe(dir, exe) {
+function copyExe(dir, exe, renamed) {
   const ext = process.platform === "win32" ? ".exe" : "";
   const src = path.join(dir, exe + ext);
-  const dest = path.join(binDir, `${exe}.exe`);
+  const dest = path.join(binDir, `${renamed ?? exe}.exe`);
 
   // For some reason, the copy operation fails in Windows CI if the file already exists.
   if (process.platform === "win32" && fs.existsSync(dest)) {
