+++ tests/analysis_tests/tests-generic-jsx-transform/package.json
@@ -2,8 +2,8 @@
   "name": "@tests/generic-jsx-transform",
   "private": true,
   "scripts": {
-    "build": "rescript",
-    "clean": "rescript clean -with-deps"
+    "build": "rescript legacy build",
+    "clean": "rescript legacy clean"
   },
   "dependencies": {
     "rescript": "workspace:^"
