+++ tests/analysis_tests/tests-incremental-typechecking/package.json
@@ -2,8 +2,8 @@
   "name": "@tests/incremental-typechecking",
   "private": true,
   "scripts": {
-    "build": "rescript",
-    "clean": "rescript clean -with-deps"
+    "build": "rescript legacy build",
+    "clean": "rescript legacy clean"
   },
   "dependencies": {
     "rescript": "workspace:^"
