+++ tests/analysis_tests/tests-reanalyze/deadcode/package.json
@@ -2,8 +2,8 @@
   "name": "@tests/reanalyze-deadcode",
   "private": true,
   "scripts": {
-    "build": "rescript",
-    "clean": "rescript clean -with-deps"
+    "build": "rescript legacy build",
+    "clean": "rescript legacy clean"
   },
   "dependencies": {
     "@rescript/react": "link:../../../dependencies/rescript-react",
