+++ tests/analysis_tests/tests-reanalyze/termination/package.json
@@ -2,8 +2,8 @@
   "name": "@tests/reanalyze-termination",
   "private": true,
   "scripts": {
-    "build": "rescript",
-    "clean": "rescript clean -with-deps"
+    "build": "rescript legacy build",
+    "clean": "rescript legacy clean"
   },
   "dependencies": {
     "rescript": "workspace:^"
