+++ tests/docstring_tests/package.json
@@ -3,8 +3,8 @@
   "type": "module",
   "private": true,
   "scripts": {
-    "build": "rescript",
-    "clean": "rescript clean -with-deps",
+    "build": "rescript legacy build",
+    "clean": "rescript legacy clean",
     "dev": "rescript -w"
   },
   "dependencies": {
